+++ roach/string/bin.py
@@ -6,6 +6,7 @@ import struct
 
 from roach.string.ops import Padding, hex, unhex
 
+
 class IntWorker(object):
     fmt = None
     size = None
@@ -14,7 +15,7 @@ class IntWorker(object):
         self.mul = mul
 
     def __call__(self, value):
-        if isinstance(value, (int, long)):
+        if isinstance(value, (int, float)):
             return struct.pack(self.fmt, value)
 
         count = len(value) / self.size
@@ -22,44 +23,53 @@ class IntWorker(object):
         if not count:
             return
 
-        ret = struct.unpack(self.fmt*count, value)
+        ret = struct.unpack(self.fmt * count, value)
         return ret[0] if count == 1 else ret
 
     def __mul__(self, other):
         """Helper method for roach.structure.Structure."""
         return self.__class__(other)
 
+
 class Int8(IntWorker):
     fmt = "b"
     size = 1
+
 
 class UInt8(IntWorker):
     fmt = "B"
     size = 1
 
+
 class Int16(IntWorker):
     fmt = "h"
     size = 2
+
 
 class UInt16(IntWorker):
     fmt = "H"
     size = 2
 
+
 class Int32(IntWorker):
     fmt = "i"
     size = 4
+
 
 class UInt32(IntWorker):
     fmt = "I"
     size = 4
 
+
 class Int64(IntWorker):
     fmt = "q"
     size = 8
 
+
 class UInt64(IntWorker):
     fmt = "Q"
     size = 8
+
 
 int8 = Int8()
 uint8 = UInt8()
@@ -70,14 +80,16 @@ uint32 = UInt32()
 int64 = Int64()
 uint64 = UInt64()
 
+
 def bigint(s, bitsize):
-    if isinstance(s, (int, long)):
+    if isinstance(s, (int, float)):
         return Padding.null(unhex("%x" % s)[::-1], bitsize / 8)
 
     if len(s) < bitsize / 8:
         return
 
     return int(hex(s[:bitsize / 8][::-1]), 16)
+
 
 # TODO Do we need any love on top of this?
 unpack = struct.unpack
